Website. DB interaction script.
The website has the tiniest amount of functionality, ever. It's just a
little textbox that adds something to the database. db.py is both a
manager for the (relatively small) amount of sqlalchemy boilerplate, and
a little script that lets you do some basic functionality without having
to write up a python script, or run the REPL. It's pretty nice.
Eventually I want to have it take command line arguments as well.